Implement "HUD-Hooking"
A lot of mods nowadays love to edit parts of the base game to add features or expand on portions of the game, but huds are prohibitively inaccessible and mods have to cope by just drawing parts of or the entirety of the same hud again.
This can be mitigated by allowing mods to "hook" into the original hud functions and tell them to not run, thereby increasing performance of the game and allowing for more extensive hud edits. The table as of now only consists of groups of huds where all elements are local and inaccessible normally. More could be added in the future if needed.