Implement more validation when choosing subs and skillcards for shuffletime
Skillcards will check to ensure a skill definition has a valid entry in the attackDefs table, if not the game will print a warning and return an agi skillcard.

Subpersonas will check to ensure that a subpersona definition has a valid entry in the subpersonaList table, and that a number of values are defined. If any of the checks fail, the game will print a warning and return nil.