Add Fallback graphic for cutins & support super cutins
We were falling back onto CUTIN_D if somehow an mo's hudcutin value became nil when the cutin_face_prefix value is being assigned. This is an extremely unlikely scenario compared to the more common instance of mo.hudcutin being defined, but not pointing to a valid sprite, so this functionality went unused.

This commit brings this functionality back by adding an else case to the v.patchExists check so CUTIN_D is used when the desired graphic doesnt exist. We also make sure we only animate three frames of a character's cutin, the else case exposed how the game was actually looking to draw the player cutin for four frames instead of just three(All vanilla characters only have three frames to their cutins).

While we are here, we can add support for super cutins, so cutins have parity with hudbusts and hudaoas.