Explicitly check for nil for integer values on subs + debug logging
Mistake on my part, I believed the not condition would be checking for just a value of nil but it also trips when a value is 0, which is common for subs which use the sprite frame A and in other situations.

The validation now checks explicitly for nil on the rest of the integer values, we also cache the subpersona result so we can save on direct accesses.

A set of debug prints have also been added so modders with the setting enabled can check to see how the game parsed the problematic sub's values.