Fix Modding Hooks only sending one argument to functions
Apparently, when you encapsulate an unpack call on a table with more than one element within parenthesis, lua will, for whatever reason, only return one element instead of all of them.
This was causing a number of Hooks that were being ran with more than one argument to be inoperable.
Removing the parenthesis from the unpack call fixes this issue.