Add a limiter to amount of rendered dungeon logs.
One place of potential performance degradation is when the game has to render a lot of logs, usually when the length of the array causes the newest elements to fly off the screen. Adding a limiter here helps with that, and can be adjusted as needed(maybe with a user accessible option in the future?). For this commit the limiter is at 7 logs at any given point.