Give control back to players right when COND_RAGE wears off
COND_RAGE had an interesting quirk where if the effect weared off on its own, the affected character would still autoattack on the turn that the status effect was cleared.  This has been changed so that when the cureStatus function is called, we also turn t_forcecontrol off and exit the function so the character can act as normal.