Prevent "Exit Spamming" on certain menus in Pause_MP
One problem with the dungeon menu is that the game can "reset" the closing animation if you press the exit button while the menu is exiting.  This can be annoying when you are in a situation where you need to close the menu quickly to avoid an enemy.  This has been remedied with the new mo variable "m_exitmenu" All relevant menus will now set, and check for this value when processing BTN_USE to prevent the closing animation from being reset.