Play a confirmation sound when a CT_CVARTEXT option entry is registered
There's no confirmation to the user when they send a text value to a CT_CVARTEXT option.  This has caused a lot of people to not understand the proper way to change their name in the Player Setup menu.  This commit plays the confirmation sound when you press enter, to let the user know their input has been submitted.