Implement Spectrum's "chatput" fix
With the jump to 2.2.10, we started having issues with inputs to the chat box leaking out onto general gameplay, causing annoyances. This implements Spectrum's fix to ensure we are not leaking inputs to gameplay when we are typing in chat. The fix was expanded to block arrow key inputs so moving up and down the chatlog wouldn't move menus.