Allow Custom Dungeons to overwrite equipment lab text
I wanted to originally keep this feature as just a part of a mod until I realized that shopkeeper mods likely completely overhaul the equipment lab code to replace the text.
This commit rawsets the text data to simplify the process for Shopkeeper mods to edit it, and also makes the equipment lab check for a new custom dungeon table before assigning anything to the now empty local textdata table.