Add BTL_initSkillChange and BTL_changeSkillSelect to Battle.LUA
Apparently the web IDE is cursed, so I extracted the changes made in !48 and rehosted them here so the changes can be viewed(R.E.D agreed to let me do this).

This allows skills to be changed into other skills (instead of editing the attackDefs of the skill, which causes desynchs).

Authored by: @redmetalsonic