Add unique case for null passives in status hitrate calculation
Assuming I correctly understand the math behind this. An entity with a Null Status Passive currently has a 1/101 chance of being inflicted with the passive(The 1/101 comes from 0 being included in the range). Since this likely is not the intended mechanics, status_hitrate now has a special case where it defaults to -1 if status_resist_mult is 100, providing a true 0% hitrate when a null status passive is present.