Respect C array boundaries of skin array
We can only index the skins array between the values of 0 and 254. This was only found out after someone possibly set their defaultskin cv value to a mod character, then restarted the game with said mod unloaded, then went into the player setup menu. The game iterated through the whole skin table trying to find the non-existant skin and reached index 255 which is outside the bounds.